--- src/main/java/com/campus/campus/global/util/jwt/logout/presentation/LogoutResponseCode.java
@@ -1,18 +1,0 @@
-package com.campus.campus.global.util.jwt.logout.presentation;
-
-import org.springframework.http.HttpStatus;
-
-import com.campus.campus.global.common.response.ResponseCodeInterface;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum LogoutResponseCode implements ResponseCodeInterface {
-	LOGOUT_SUCCESS(200, HttpStatus.OK, "로그아웃에 성공했습니다.");
-
-	private final int code;
-	private final HttpStatus status;
-	private final String message;
-}
